requirements.txt
selenium
tqdm
pylint
black
scikit-image
opencv-python
tensorflow
keras
Pillow
cython
h5py
imgaug
IPython[all]
wandb
typer[all]
ultralytics
torch
numpy
matplotlib
